requirements.txt
python-telegram-bot==21.5
sentence-transformers==3.0.1
numpy==1.26.4
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.12.5